pyproject.toml
[project]
name = "zuul-capacity"
version = "0.3.0"
description = "Zuul capacity exporter"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openstacksdk>=4.1.0",
    "prometheus-client>=0.21.0",
    "pyyaml>=6.0.2",
]
